+++ Backend/src/GameTest.Application/Features/Catalog/ReadModels/PassiveAbilityReadModel.cs
@@ -6,7 +6,7 @@ namespace GameTest.Application.Features.Catalog.ReadModels
     {
         public string Name { get; init; } = string.Empty;
         public string Description { get; init; } = string.Empty;
-        public double Bonus { get; init; }
+        public float Bonus { get; init; }
         public PassiveAbilityType Type { get; init; }
     }
 }
