+++ Backend/src/GameTest.Application/Features/PlayerProgression/ReadModels/PlayerWeaponPropertyReadModel.cs
@@ -8,9 +8,9 @@ namespace GameTest.Application.Features.PlayerProgression.ReadModels
         public int StatId { get; init; }
         public string StatName { get; init; } = null!;
         public WeaponStatType StatType { get; init; }
-        public double Value { get; init; }
+        public float Value { get; init; }
         public int Level { get; init; }
-        public double? NextLevelValue { get; init; }
+        public float? NextLevelValue { get; init; }
         public int? NextLevelPrice { get; init; }
         public int MaxLevel { get; init; }
     }
